feat!: rename lambda function with test with setting
