+++ .gitattributes
@@ -19,6 +19,6 @@
 /API.md linguist-generated
 /LICENSE linguist-generated
 /package.json linguist-generated
-/src/log-archiver-function.ts linguist-generated
+/src/funcs/log-archiver-function.ts linguist-generated
 /tsconfig.dev.json linguist-generated
 /yarn.lock linguist-generated
