+++ .projen/files.json
@@ -14,7 +14,7 @@
     ".projen/files.json",
     ".projen/tasks.json",
     "LICENSE",
-    "src/log-archiver-function.ts",
+    "src/funcs/log-archiver-function.ts",
     "tsconfig.dev.json"
   ],
   "//": "~~ Generated by projen. To modify, edit .projenrc.js and run \"npx projen\"."
