+++ package.json
@@ -9,8 +9,8 @@
     "build": "npx projen build",
     "bump": "npx projen bump",
     "bundle": "npx projen bundle",
-    "bundle:log-archiver.lambda": "npx projen bundle:log-archiver.lambda",
-    "bundle:log-archiver.lambda:watch": "npx projen bundle:log-archiver.lambda:watch",
+    "bundle:funcs/log-archiver.lambda": "npx projen bundle:funcs/log-archiver.lambda",
+    "bundle:funcs/log-archiver.lambda:watch": "npx projen bundle:funcs/log-archiver.lambda:watch",
     "clobber": "npx projen clobber",
     "compat": "npx projen compat",
     "compile": "npx projen compile",
